implement stubs for addressspace, category and organisation GraphQL queries.
